Task 3/push/requirements.txt
Flask==2.3.2
gunicorn==20.1.0
google-cloud-bigquery==3.11.2